requirements.txt
streamlit
firebase-admin
pandas
requests
python-dotenv
google-generativeai
anthropic
